Move `<Card />` to `card/` subfolder and add  new`index.js` file to export all card components
