+++ packages/components/src/card/stories/_index.js
@@ -6,7 +6,7 @@ import { text, boolean } from '@storybook/addon-knobs';
 /**
  * Internal dependencies
  */
-import Card from '../index';
+import { Card } from '../';
 import CardBody from '../body';
 import CardFooter from '../footer';
 import CardHeader from '../header';
